Rename handlers folders
