+++ social/handlers_github/base.py
@@ -1,4 +1,3 @@
-import requests
 import logging
 import re
 from typing import Any, Callable
