+++ social/handlers_github/profcomff_issues.py
@@ -20,7 +20,7 @@ def issue_opened(event):
     """
     logger.debug("Issue %s created (node_id=%s)", event["issue"].get("url"), event["issue"].get("node_id"))
     r = github.request_gql(
-        'social/github/profcomff_issues.gql',
+        'social/handlers_github/profcomff_issues.gql',
         'AddToScrum',
         projectId=PROJECT_NODE_ID,
         contentId=event["issue"].get("node_id")
